Fix prev/next file commands
The editor glitched when prev/next file shortcut or command was selected. There were some strange things happening with these for loops (id being string and not int) and later it would get fuckered in the project.js when trying to add or subtract 1 to move between files. This should fix it.

(Apparently iterating arrays with for..in is not recommended...)